% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistMatrix.R
\name{calcDistMat}
\alias{calcDistMat}
\title{Title Calculate dissimilarity matrix (local cost matrix)}
\usage{
calcDistMat(x, y, dist.method)
}
\arguments{
\item{x}{single cell gene expression data: query trajectory (genes on rows, cells on columns)}

\item{y}{single cell gene expression data: reference trajectory (genes on rows, cells on columns)}

\item{dist.method}{distance metric (see: proxy)}
}
\value{
dissimilarity matrix (local cost matrix) between x and y
}
\description{
Title Calculate dissimilarity matrix (local cost matrix)
}
\examples{
calcDistMat(x,y)
}
