% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLocalMinima.R
\name{findGlobalMinima}
\alias{findGlobalMinima}
\title{This function calculates the global minima of the cost matrix;
These minima will be the potential starting points
The minima calculation is performed on the rows of the lm matrix:
i.e. for each row, it finds the local minima, and also checks the edges.}
\usage{
findGlobalMinima(lm, costMatrix)
}
\description{
This function calculates the global minima of the cost matrix;
These minima will be the potential starting points
The minima calculation is performed on the rows of the lm matrix:
i.e. for each row, it finds the local minima, and also checks the edges.
}
