% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLocalMinima.R
\name{findLocalMinima}
\alias{findLocalMinima}
\title{This function calculates the local minima of the lm matrix;
These minima will be the potential starting points
The minima calculation is performed on the rows of the lm matrix:
i.e. for each row, it finds the local minima, and also checks the edges.}
\usage{
findLocalMinima(lm, similarityThresh)
}
\description{
This function calculates the local minima of the lm matrix;
These minima will be the potential starting points
The minima calculation is performed on the rows of the lm matrix:
i.e. for each row, it finds the local minima, and also checks the edges.
}
