% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySub2Ind.R
\name{ind2sub}
\alias{ind2sub}
\title{Title Convert linear indices to a row-columns matrix}
\usage{
ind2sub(rowNum, ind)
}
\arguments{
\item{rowNum}{number of rows in the desired matrix}

\item{ind}{a vector of linear indices}
}
\value{
a data.frame with the row and columns odf the linear indices
}
\description{
Title Convert linear indices to a row-columns matrix
}
\examples{
ind2sub(rowNum, ind)
}
