% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interWeights.R
\name{interWeights}
\alias{interWeights}
\title{Title interpolation of the expression data along one trajectory}
\usage{
interWeights(expDataBatch, trajCond, winSz, numPts)
}
\arguments{
\item{expDataBatch}{raw single cell gene expression data (genes on rows, cells on columns)}

\item{trajCond}{a vector of pseudo-time scores of the data-points whose length equals to the number of samples}

\item{winSz}{window size of the interpolation}

\item{numPts}{number of desired interpolated points}
}
\value{
a list of 3: interpolatedVals, error and pseudo-time scores of the interpolated points;
interpolatedVals: a matrix of the interpolated points (genes*numPts)
error: a matrix of interpolated error for each interpolated data-point (genes*numPts)
traj: pseudo-time scores of the interpolated points
}
\description{
Title interpolation of the expression data along one trajectory
}
