% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localAlignAutoThreshold.R
\name{localAlignAutoThreshold}
\alias{localAlignAutoThreshold}
\title{Title Performs local alignment between two single-cell gene-expression patterns under automatically set threshold}
\usage{
localAlignAutoThreshold(x, y = NULL, dist.method = "Euclidean",
  step.pattern = symmetric2, distance.only = F, normDist = T, ...)
}
\arguments{
\item{x}{query data matrix: genes on rows and cells on columns OR local cost matrix}

\item{y}{reference data matrix: genes on rows and cells on columns, unused if x is
given as a cost matrix}

\item{dist.method}{distance metric, used by the function "dist" in the package 'proxy'}

\item{step.pattern}{a stepPattern object describing the local warping steps allowed with their cost}

\item{distance.only}{only compute distance (no backtrack, faster)}

\item{normDist}{logical; scale distance matrix?}

\item{...}{}
}
\value{
local alignment (as in localAlign function); plot of the picked threshold
}
\description{
Title Performs local alignment between two single-cell gene-expression patterns under automatically set threshold
}
\examples{
localAlignAutoThreshold = function(x, y)
}
