% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{plotAlign}
\alias{plotAlign}
\title{Title Plot the alignment over the dissimilarity matrix}
\usage{
plotAlign(alignment)
}
\arguments{
\item{alignment}{output from the localAlign/globalAlign functions}
}
\value{
NA
}
\description{
Title Plot the alignment over the dissimilarity matrix
}
\examples{
plotAlign(globalAlignment)
}
