% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingActualData.R
\name{plotMapping}
\alias{plotMapping}
\title{Title Plot the alignment-based mapping between real cells from query and reference trajectories}
\usage{
plotMapping(mapping)
}
\arguments{
\item{mapping}{mapping output from the mapRealDataGlobal or mapRealDataLocal}
}
\value{
plot
}
\description{
Title Plot the alignment-based mapping between real cells from query and reference trajectories
}
\examples{
plotMapping(mapping)
}
