% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotimeClust.R
\name{pseudotimeClust}
\alias{pseudotimeClust}
\title{Title cluster genes based on their pseudotime shifts}
\usage{
pseudotimeClust(x, y, k = 2)
}
\arguments{
\item{x}{interpolated scaled expression of n genes (rows) in condition 1}

\item{y}{interpolated scaled expression of n genes (rows) in condition 2}

\item{k}{number of clusters}
}
\value{
clustering result
}
\description{
Title cluster genes based on their pseudotime shifts
}
\examples{
pseudotimeClust(x, y, k)
}
