% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleInterpolate.R
\name{scaleInterpolate}
\alias{scaleInterpolate}
\title{Title Scale interpolated data to [0,1]}
\usage{
scaleInterpolate(interDataConds)
}
\arguments{
\item{interDataConds}{interpolated data; output from the function: interWeights}
}
\value{
a list consisting from the interpolated-scaled values, interpolated-scaled error and pseudo-time scores of the interpolated points
}
\description{
Title Scale interpolated data to [0,1]
}
\examples{
scaleInterpolate(interDataConds)
}
