% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySub2Ind.R
\name{sub2ind}
\alias{sub2ind}
\title{Title Convert a row-column vector to a linear index}
\usage{
sub2ind(rowNum, rows, cols)
}
\arguments{
\item{rowNum}{nuber of rows in the data matrix}

\item{rows}{a vector containing the rows of the converted items}

\item{cols}{a vector containing the columns of the converted items}
}
\value{
a vector of the linear indices
}
\description{
Title Convert a row-column vector to a linear index
}
\examples{
sub2ind(rowNum, rows, cols)
}
